move NewID() to utils package
