+++ server/utils/utils.go
@@ -1,10 +1,13 @@
 package utils
 
 import (
+	"bytes"
+	"encoding/base32"
 	"regexp"
 
 	"github.com/jfrerich/mattermost-plugin-bookmarks/server/pluginapi"
 	"github.com/mattermost/mattermost-server/v5/plugin"
+	"github.com/pborman/uuid"
 )
 
 // GetPostIDFromLink extracts a PostID from a link
@@ -34,4 +37,18 @@ func GetLegendText() string {
 // getSiteURL returns the SiteURL from the config settings
 func GetSiteURL(api pluginapi.API) string {
 	return *api.GetConfig().ServiceSettings.SiteURL
+}
+
+var encoding = base32.NewEncoding("ybndrfg8ejkmcpqxot1uwisza345h769")
+
+// NewID is a globally unique identifier.  It is a [A-Z0-9] string 26
+// characters long.  It is a UUID version 4 Guid that is zbased32 encoded
+// with the padding stripped off.
+func NewID() string {
+	var b bytes.Buffer
+	encoder := base32.NewEncoder(encoding, &b)
+	_, _ = encoder.Write(uuid.NewRandom())
+	encoder.Close()
+	b.Truncate(26) // removes the '==' padding
+	return b.String()
 }
